Finally have a fully working version of pause time. It has inbuilt limitations required by this system:  * No shooting people into pause time. More generally no interacting in any way with paused guys whatsoever.
As a further reminder: FUTURE SELVES, DON"T IMPLEMENT MORE PAUSE TIME MANIPULATION.

Shooting people into pause time would most likely break the assertion that there is always 1 player frame as remote departure edits take time to propagate through yet do not change relative index.

Anyway have a fit with this version and test bugs, many things are untested because I can't be bothered writing replays to test it. Hold 'P' to test the engine is still in this version. Removing it soon to work on other things.

+++ trunk/Guy.h
@@ -19,7 +19,7 @@ namespace hg {
             int boxPauseLevel,
             hg::TimeDirection nTimeDirection,
             int pauseLevel,
-            size_t nRelativeIndex,
+            int nRelativeIndex,
             int nSubimage);
 
         ~Guy();
@@ -42,7 +42,7 @@ namespace hg {
 
         TimeDirection getTimeDirection() const {return data->timeDirection;}
         int getPauseLevel() const {return data->pauseLevel;}
-        size_t getRelativeIndex() const {return data->relativeIndex;}
+        int getRelativeIndex() const {return data->relativeIndex;}
         int getSubimage() const {return data->subimage;}
 
         const static int animationLength = 13;
@@ -74,7 +74,7 @@ namespace hg {
 
                  TimeDirection ntimeDirection,
                  int npauseLevel,
-                 size_t nrelativeIndex,
+                 int nrelativeIndex,
 
                  int nsubimage) :
             x(nx),
@@ -113,7 +113,7 @@ namespace hg {
 
             TimeDirection timeDirection;
             int pauseLevel;
-            size_t relativeIndex;
+            int relativeIndex;
 
             int subimage;
         };
